Adds a number of performance tweaks to the loading process. This brings down the load for fishtornado on a Quest2 headset from ~30-40s to ~5-6s.