@Slumpwuffle 2023-05-02
Okay, github doesn't let me reply to your comment directly, but:
>Not sure. Is this change alone actually causing a notable perf increase? Why is that better?
In the first case, you have a cmp+jmp in the loop. But in the new code you have a load+call.
Because those are indirect calls, I'm guessing the compiler cannot go full-naked function, and still has to respect some calling conventions, hence may move stuff to specific registers, save the stack, etc. While the previous code didn't have any of that.

The answer is that branches are indeed slower. Admittedly most of the gains are from memory allocation, but branches require pipeline and cache flushes if mispredicted, and often pipeline stalls even if not. Static flows are almost always faster even with the additional memory access and instructions because those will end up in cache. On what was a 7s load, the branch removal was roughly a 1s gain, the memory allocs were the rest of the 6s.