@Keenuts 2023-05-03
> The answer is that branches are indeed slower. Admittedly most of the gains are from memory allocation, but branches require pipeline flushes if mispredicted or lacking prediction, and often pipeline stalls even if not. Static flows are almost always faster even with the additional memory access and instructions because those will end up in cache. On what was a 7s load, the branch removal was roughly a 1s gain, the memory allocs were the rest of the 6s.

So you have measurements to back this? Branch prediction can indeed be wrong, so is speculative execution. (Since we are not at a ring boundary, I believe speculative execution is still authorized and unhindered by all the spectre mitigations and such).
Anyway, I'm definitely not an expert of micro-architectural details as they are way too complex nowadays.

What I did on my end was to keep all the PR changes (including the pre-allocation from "Speed up obj-based mesh loading"), added some timers and did multiple headless runs through perf  with 0 frames on 15_basic_material. What I compared was the PR as it is, and PR without the branch fix. Optimization enabled of course, as comparing without wouldn't make much sense.
It's on my workstation, so obviously some noise, so added the margins I observed.

```
With the branch fix:
page_faults  : average=54259         (+- 20662)
icache_miss  : average=1.2M          (+- 184K)
cache_miss   : average=199M          (+- 26M)
branch_miss  : average=65M           (+- 1M)
branch       : average=4B            (+- 1M)
instruction  : average=26B           (+- 4M)
model        : average=0.896s        (+- 0.020s)
texture      : average=2.742s        (+- 0.106s)

Without the branch fix:
page_faults  : average=55768         (+- 10337)
icache_miss  : average=636K          (+- 22K)
cache_miss   : average=16M           (+- 3M)
branch_miss  : average=65M           (+- 815K)
branch       : average=4B            (+- 527K)
instruction  : average=22B           (+- 3M)
model        : average=0.102         (+- 0.007s)
texture      : average=2.739         (+- 0.070s)
```

(Raw results in case you want to have more precision)
```
With branch fix (pr as is):
page_faults  : average=54259.083	(margin= 20662.000)
icache_miss  : average=1180922.500	(margin=184062.000)
cache_miss   : average=199257846.667	(margin=26665308.000)
branch_miss  : average=65546041.750	(margin=976332.000)
branch       : average=4596896371.667	(margin=907387.000)
instruction  : average=26363841984.500	(margin=4253171.000)
model        : average=0.896	(margin=     0.020)
texture      : average=2.742	(margin=     0.106)

Without branch fix (only smarter allocations):
page_faults  : average=55768.833	(margin= 10337.000)
icache_miss  : average=636745.833	(margin= 22821.000)
cache_miss   : average=16867040.500	(margin=3138835.000)
branch_miss  : average=65119469.583	(margin=815188.000)
branch       : average=4234112410.167	(margin=527591.000)
instruction  : average=22350433883.250	(margin=3907268.000)
model        : average=0.102	(margin=     0.007)
texture      : average=2.739	(margin=     0.070)
```