@chaoticbob 2023-05-03
If there is a noticable performance improvement for Android, I'd like to get some of these changes in soon and worry about the specifics of the benchmarking later (assuming they're not egregious). vk_15_basic_materials takes minutes to load currently on Android and I'd like for the Android fixes to not conflict with this PR.