@Slumpwuffle 2023-05-03
Yep, okay, I agree, it's not the branching. I only saw the difference in one run before, and it could've been inconsistent alloc behavior. Trying it now isolating the branching vs. functional variants, there's no meaningful net impact (seeing ~0.048-0.049 seconds for loading the shark each time regardless of approach). I think maybe losing the branches is faster on average by about ~0.001s, but it's likely not worth the code churn so I'll revert it.